requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
click==8.3.0
fastapi==0.118.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
psycopg==3.2.10
psycopg-c==3.2.10
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
